requirements.txt
google-cloud-aiplatform
google-cloud-storage
google-genai
gitpython
google-adk
google-adk[eval]
vertexai
PyMuPDF
Pillow